+++ examples/binary/main.tf
@@ -5,21 +5,13 @@ resource "random_string" "random" {
 
 module "binary_string" {
   source        = "./../../"
-  name          = "example-binary-secret"
-  description   = "a binary secret example"
+  name          = var.name
+  description   = var.secret_description
   secret_policy = local.policy
   secrets = {
     secret1 = {
       secret_binary = base64encode(random_string.random.result)
     }
   }
-  tags = {
-    Environment        = "examples"
-    "user::CostCenter" = "terraform-registry"
-    department         = "DevOps"
-    Project            = "Examples"
-    Owner              = "Boldlink"
-    LayerName          = "cExample"
-    LayerId            = "cExample"
-  }
+  tags = var.tags
 }