+++ examples/complete/locals.tf
@@ -2,19 +2,9 @@ locals {
   account_id       = data.aws_caller_identity.current.account_id
   partition        = data.aws_partition.current.partition
   region           = data.aws_region.current.name
-  name             = "example-complete-secret"
-  filename         = "mysql-lambda.zip"
-  cidr_block       = "192.168.0.0/16"
-  tag_env          = "dev"
-  rotation_subnet1 = cidrsubnet(local.cidr_block, 7, 50)
-  rotation_subnet2 = cidrsubnet(local.cidr_block, 7, 60)
-  rotation_subnet3 = cidrsubnet(local.cidr_block, 7, 70)
-  rotation_subnets = [local.rotation_subnet1, local.rotation_subnet2, local.rotation_subnet3]
-
-  az1 = data.aws_availability_zones.available.names[0]
-  az2 = data.aws_availability_zones.available.names[1]
-  az3 = data.aws_availability_zones.available.names[2]
-  azs = [local.az1, local.az2, local.az3]
+  private_subnets  = [cidrsubnet(var.cidr_block, 8, 1), cidrsubnet(var.cidr_block, 8, 2), cidrsubnet(var.cidr_block, 8, 3)]
+  isolated_subnets = [cidrsubnet(var.cidr_block, 8, 10), cidrsubnet(var.cidr_block, 8, 11), cidrsubnet(var.cidr_block, 8, 13)]
+  azs              = [data.aws_availability_zones.available.names[0], data.aws_availability_zones.available.names[1], data.aws_availability_zones.available.names[2]]
   policy = jsonencode(
     {
       Version = "2012-10-17",
@@ -33,17 +23,6 @@ locals {
   lambda_policy = jsonencode({
     Version = "2012-10-17"
     Statement = [
-      {
-        Sid    = "GiveSpecificNetworkInterfacePermissions",
-        Effect = "Allow",
-        Action = [
-          "ec2:CreateNetworkInterface",
-          "ec2:DeleteNetworkInterface",
-          "ec2:DescribeNetworkInterfaces",
-          "ec2:DetachNetworkInterface",
-        ],
-        Resource = ["*"]
-      },
       {
         Sid    = "GiveSpecificSecretPermissions",
         Effect = "Allow",
@@ -64,13 +43,4 @@ locals {
         Resource = ["*"]
     }]
   })
-  tags = {
-    Environment        = "examples"
-    "user::CostCenter" = "terraform-registry"
-    department         = "DevOps"
-    Project            = "Examples"
-    Owner              = "Boldlink"
-    LayerName          = "cExample"
-    LayerId            = "cExample"
-  }
 }