+++ examples/complete/mysql_db.tf
@@ -1,44 +1,36 @@
-module "rds_kms" {
-  source              = "boldlink/kms/aws"
-  version             = "1.1.0"
-  description         = "key for mysql"
-  create_kms_alias    = true
-  enable_key_rotation = true
-  alias_name          = "alias/${local.name}-rds-key"
-  tags                = local.tags
+resource "random_password" "mysql_password" {
+  length  = var.length
+  special = var.special
 }
 
-resource "random_password" "mysql_password" {
-  length  = 16
-  special = false
-}
 
 module "mysql" {
   source                              = "boldlink/rds/aws"
   version                             = "1.1.2"
-  engine                              = "mysql"
-  instance_class                      = "db.t3.micro"
-  subnet_ids                          = flatten(module.rotation_vpc.private_subnet_id)
-  name                                = "exampledb"
-  username                            = "admin"
+  engine                              = var.engine
+  instance_class                      = var.instance_class
+  subnet_ids                          = flatten(module.vpc.isolated_subnet_id)
+  name                                = var.db_name
+  username                            = var.username
   password                            = random_password.mysql_password.result
-  kms_key_id                          = module.rds_kms.arn
-  port                                = 3306
-  iam_database_authentication_enabled = true
-  multi_az                            = true
-  create_security_group               = true
-  enabled_cloudwatch_logs_exports     = ["general", "error", "slowquery"]
-  create_monitoring_role              = true
-  monitoring_interval                 = 30
-  deletion_protection                 = false
-  vpc_id                              = module.rotation_vpc.id
+  kms_key_id                          = module.kms.arn
+  port                                = var.port
+  iam_database_authentication_enabled = var.iam_database_authentication_enabled
+  multi_az                            = var.multi_az
+  create_security_group               = var.create_security_group
+  enabled_cloudwatch_logs_exports     = var.enabled_cloudwatch_logs_exports
+  create_monitoring_role              = var.create_monitoring_role
+  monitoring_interval                 = var.monitoring_interval
+  deletion_protection                 = var.deletion_protection
+  vpc_id                              = module.vpc.id
   assume_role_policy                  = data.aws_iam_policy_document.monitoring.json
   policy_arn                          = "arn:${local.partition}:iam::aws:policy/service-role/AmazonRDSEnhancedMonitoringRole"
-  major_engine_version                = "8.0"
+  major_engine_version                = var.major_engine_version
 
   tags = merge({
-    "InstanceScheduler" = true },
-  local.tags)
+    InstanceScheduler = true },
+    { Name = var.name },
+  var.tags)
 
   security_group_ingress = [
     {
@@ -46,7 +38,7 @@ module "mysql" {
       from_port   = 0
       to_port     = 0
       protocol    = "-1"
-      cidr_blocks = [local.cidr_block]
+      cidr_blocks = local.private_subnets
     }
   ]
   security_group_egress = [
@@ -55,7 +47,7 @@ module "mysql" {
       from_port   = 0
       to_port     = 0
       protocol    = -1
-      cidr_blocks = [local.cidr_block]
+      cidr_blocks = local.private_subnets
     }
   ]
 }