+++ examples/key-value/main.tf
@@ -1,7 +1,7 @@
 module "key_pair_string" {
   source        = "./../../"
-  name          = "keyvalue-secret-example"
-  description   = "this is a key-value secret example"
+  name          = var.name
+  description   = var.secret_description
   secret_policy = local.policy
   secrets = {
     key_value1 = {
@@ -13,13 +13,5 @@ module "key_pair_string" {
       )
     }
   }
-  tags = {
-    Environment        = "examples"
-    "user::CostCenter" = "terraform-registry"
-    department         = "DevOps"
-    Project            = "Examples"
-    Owner              = "Boldlink"
-    LayerName          = "cExample"
-    LayerId            = "cExample"
-  }
+  tags = var.tags
 }