+++ examples/minimum/main.tf
@@ -1,19 +1,11 @@
 module "minimum" {
   source      = "./../../"
-  name        = "example-minimum-secret"
-  description = "Example minimum secret"
+  name        = var.name
+  description = var.secret_description
   secrets = {
     single_string = {
       secret_string = "Sample String"
     }
   }
-  tags = {
-    Environment        = "examples"
-    "user::CostCenter" = "terraform-registry"
-    department         = "DevOps"
-    Project            = "Examples"
-    Owner              = "Boldlink"
-    LayerName          = "cExample"
-    LayerId            = "cExample"
-  }
+  tags = var.tags
 }